+++ tencentcloud/essbasic/v20210526/models.go
@@ -5223,6 +5223,66 @@ func (r *ChannelModifyRoleResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
 }
 
+type ChannelOrganizationInfo struct {
+	// 电子签企业Id
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	OrganizationId *string `json:"OrganizationId,omitnil" name:"OrganizationId"`
+
+	// 电子签企业OpenId
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	OrganizationOpenId *string `json:"OrganizationOpenId,omitnil" name:"OrganizationOpenId"`
+
+	// 企业名称
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	OrganizationName *string `json:"OrganizationName,omitnil" name:"OrganizationName"`
+
+	// 企业信用代码
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	UnifiedSocialCreditCode *string `json:"UnifiedSocialCreditCode,omitnil" name:"UnifiedSocialCreditCode"`
+
+	// 法人姓名
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	LegalName *string `json:"LegalName,omitnil" name:"LegalName"`
+
+	// 法人OpenId
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	LegalOpenId *string `json:"LegalOpenId,omitnil" name:"LegalOpenId"`
+
+	// 超管姓名
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	AdminName *string `json:"AdminName,omitnil" name:"AdminName"`
+
+	// 超管OpenId
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	AdminOpenId *string `json:"AdminOpenId,omitnil" name:"AdminOpenId"`
+
+	// 超管手机号，脱敏后返回
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	AdminMobile *string `json:"AdminMobile,omitnil" name:"AdminMobile"`
+
+	// 企业认证状态字段。值如下：
+	// <ul>
+	//   <li>**"UNVERIFIED"**： 未认证的企业</li>
+	//   <li>**"VERIFYINGLEGALPENDINGAUTHORIZATION"**： 认证中待法人授权的企业</li>
+	//   <li>**"VERIFYINGAUTHORIZATIONFILEPENDING"**： 认证中授权书审核中的企业</li>
+	//   <li>**"VERIFYINGAUTHORIZATIONFILEREJECT"**： 认证中授权书已驳回的企业</li>
+	//   <li>**"VERIFYING"**： 认证中的企业</li>
+	//   <li>**"VERIFIED"**： 已认证的企业</li>
+	// </ul>
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	AuthorizationStatus *string `json:"AuthorizationStatus,omitnil" name:"AuthorizationStatus"`
+
+	// 企业认证方式字段。值如下：
+	// <ul>
+	//   <li>**"AuthorizationInit"**： 暂未选择授权方式</li>
+	//   <li>**"AuthorizationFile"**： 授权书</li>
+	//   <li>**"AuthorizationLegalPerson"**： 法人授权超管</li>
+	//   <li>**"AuthorizationLegalIdentity"**： 法人直接认证</li>
+	// </ul>
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	AuthorizationType *string `json:"AuthorizationType,omitnil" name:"AuthorizationType"`
+}
+
 type ChannelRole struct {
 	// 角色ID,为32位字符串
 	// 注意：此字段可能返回 null，表示取不到有效值。
@@ -7085,6 +7145,130 @@ func (r *DescribeChannelFlowEvidenceReportResponse) ToJsonString() string {
 // FromJsonString It is highly **NOT** recommended to use this function
 // because it has no param check, nor strict type check
 func (r *DescribeChannelFlowEvidenceReportResponse) FromJsonString(s string) error {
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type DescribeChannelOrganizationsRequestParams struct {
+	// 关于渠道应用的相关信息，包括渠道应用标识、第三方平台子客企业标识及第三方平台子客企业中的员工标识等内容，您可以参阅开发者中心所提供的 Agent 结构体以获取详细定义。
+	// 
+	// 此接口下面信息必填。
+	// 
+	// 渠道应用标识: Agent.AppId
+	// 第三方平台子客企业标识: Agent.ProxyOrganizationOpenId
+	// 第三方平台子客企业中的员工标识: Agent. ProxyOperator.OpenId
+	// 第三方平台子客企业和员工必须已经经过实名认证
+	Agent *Agent `json:"Agent,omitnil" name:"Agent"`
+
+	// 指定分页每页返回的数据条数，单页最大支持 200。
+	Limit *int64 `json:"Limit,omitnil" name:"Limit"`
+
+	// 子客OrganizationOpenId，定向查询某个子客的企业数据。
+	OrganizationOpenId *string `json:"OrganizationOpenId,omitnil" name:"OrganizationOpenId"`
+
+	// 企业认证状态过滤字段。可值如下：
+	// <ul>
+	//   <li>**"UNVERIFIED"**： 未认证的企业</li>
+	//   <li>**"VERIFYINGLEGALPENDINGAUTHORIZATION"**： 认证中待法人授权的企业</li>
+	//   <li>**"VERIFYINGAUTHORIZATIONFILEPENDING"**： 认证中授权书审核中的企业</li>
+	//   <li>**"VERIFYINGAUTHORIZATIONFILEREJECT"**： 认证中授权书已驳回的企业</li>
+	//   <li>**"VERIFYING"**： 认证中的企业</li>
+	//   <li>**"VERIFIED"**： 已认证的企业</li>
+	// </ul>
+	AuthorizationStatusList []*string `json:"AuthorizationStatusList,omitnil" name:"AuthorizationStatusList"`
+
+	// 指定分页返回第几页的数据，如果不传默认返回第一页。 页码从 0 开始，即首页为 0，最大20000。
+	Offset *int64 `json:"Offset,omitnil" name:"Offset"`
+}
+
+type DescribeChannelOrganizationsRequest struct {
+	*tchttp.BaseRequest
+	
+	// 关于渠道应用的相关信息，包括渠道应用标识、第三方平台子客企业标识及第三方平台子客企业中的员工标识等内容，您可以参阅开发者中心所提供的 Agent 结构体以获取详细定义。
+	// 
+	// 此接口下面信息必填。
+	// 
+	// 渠道应用标识: Agent.AppId
+	// 第三方平台子客企业标识: Agent.ProxyOrganizationOpenId
+	// 第三方平台子客企业中的员工标识: Agent. ProxyOperator.OpenId
+	// 第三方平台子客企业和员工必须已经经过实名认证
+	Agent *Agent `json:"Agent,omitnil" name:"Agent"`
+
+	// 指定分页每页返回的数据条数，单页最大支持 200。
+	Limit *int64 `json:"Limit,omitnil" name:"Limit"`
+
+	// 子客OrganizationOpenId，定向查询某个子客的企业数据。
+	OrganizationOpenId *string `json:"OrganizationOpenId,omitnil" name:"OrganizationOpenId"`
+
+	// 企业认证状态过滤字段。可值如下：
+	// <ul>
+	//   <li>**"UNVERIFIED"**： 未认证的企业</li>
+	//   <li>**"VERIFYINGLEGALPENDINGAUTHORIZATION"**： 认证中待法人授权的企业</li>
+	//   <li>**"VERIFYINGAUTHORIZATIONFILEPENDING"**： 认证中授权书审核中的企业</li>
+	//   <li>**"VERIFYINGAUTHORIZATIONFILEREJECT"**： 认证中授权书已驳回的企业</li>
+	//   <li>**"VERIFYING"**： 认证中的企业</li>
+	//   <li>**"VERIFIED"**： 已认证的企业</li>
+	// </ul>
+	AuthorizationStatusList []*string `json:"AuthorizationStatusList,omitnil" name:"AuthorizationStatusList"`
+
+	// 指定分页返回第几页的数据，如果不传默认返回第一页。 页码从 0 开始，即首页为 0，最大20000。
+	Offset *int64 `json:"Offset,omitnil" name:"Offset"`
+}
+
+func (r *DescribeChannelOrganizationsRequest) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *DescribeChannelOrganizationsRequest) FromJsonString(s string) error {
+	f := make(map[string]interface{})
+	if err := json.Unmarshal([]byte(s), &f); err != nil {
+		return err
+	}
+	delete(f, "Agent")
+	delete(f, "Limit")
+	delete(f, "OrganizationOpenId")
+	delete(f, "AuthorizationStatusList")
+	delete(f, "Offset")
+	if len(f) > 0 {
+		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "DescribeChannelOrganizationsRequest has unknown keys!", "")
+	}
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type DescribeChannelOrganizationsResponseParams struct {
+	// 企业企业信息列表。
+	ChannelOrganizationInfos []*ChannelOrganizationInfo `json:"ChannelOrganizationInfos,omitnil" name:"ChannelOrganizationInfos"`
+
+	// 指定分页返回第几页的数据。页码从 0 开始，即首页为 0，最大20000。
+	Offset *int64 `json:"Offset,omitnil" name:"Offset"`
+
+	// 指定分页每页返回的数据条数，单页最大支持 200。
+	Limit *int64 `json:"Limit,omitnil" name:"Limit"`
+
+	// 符合条件的企业数量。
+	Total *int64 `json:"Total,omitnil" name:"Total"`
+
+	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
+	RequestId *string `json:"RequestId,omitnil" name:"RequestId"`
+}
+
+type DescribeChannelOrganizationsResponse struct {
+	*tchttp.BaseResponse
+	Response *DescribeChannelOrganizationsResponseParams `json:"Response"`
+}
+
+func (r *DescribeChannelOrganizationsResponse) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *DescribeChannelOrganizationsResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
 }